move EthFilter trait to single file
