+++ crates/client/src/rpc.rs
@@ -61,7 +61,8 @@ use self::{
         cfx::{Cfx, CfxFilter},
         debug::LocalRpc,
         eth_space::{
-            eth::{Eth, EthFilter},
+            eth::Eth,
+            eth_filter::EthFilter,
             eth_pubsub::EthPubSub,
             trace::Trace as EthTrace,
         },
