+++ crates/client/src/rpc/traits/eth_space/mod.rs
@@ -1,4 +1,5 @@
 pub mod debug;
 pub mod eth;
+pub mod eth_filter;
 pub mod eth_pubsub;
 pub mod trace;
